==> data/processing/01_data_processing_purchases.ipynb <==
## Dataset: amazon-purchases

import pandas as pd

### Import dataset

df_purchases_clean = pd.read_csv("data/raw/amazon_purchases.csv")

### Exploratory data analysis


df_purchases_clean.info()  # data type

df_purchases_clean["Survey ResponseID"]

df_purchases_clean["Quantity"].sum()

df_purchases_clean[
    "Category"
].value_counts()  # different data types within the selected column

df_purchases_clean[
    "Survey ResponseID"
].nunique()  # number of unique values within the selected column

df_purchases_clean[
    df_purchases_clean["Order Date"].str.startswith("2024")
]  # filters the rows where the 'Order Date' starts with '2024'

df_purchases_clean[
    df_purchases_clean["Quantity"] == 55
]  # filters the rows where the 'Quantity' is equal '55'

#### Memory usage dataset


memory = df_purchases_clean.memory_usage(deep=True)  # RAM memory used by each column
total_memory_gb = memory.sum() / (1024**3)  # transform into GB
print(f"Total memory: {total_memory_gb:.2f} GB")

memory_percent = memory / memory.sum() * 100  # RAM memory used by each column in GB
print("\nMemory by column:")
print(memory_percent)

### Data cleaning


# Edit "Category" column and convert to "category"
def capitalized_case(s):
    if pd.isna(s):
        return s
    words = s.split("_")
    return " ".join([words[0].capitalize()] + [word.lower() for word in words[1:]])


df_purchases_clean["Category"] = (
    df_purchases_clean["Category"].apply(capitalized_case).astype("category")
)

### Reduce dataset size

# Convert columns to "category"
df_purchases_clean["Shipping Address State"] = df_purchases_clean[
    "Shipping Address State"
].astype("category")

df_purchases_clean["Survey ResponseID"] = df_purchases_clean[
    "Survey ResponseID"
].astype("category")

# Convert columns to "date"
df_purchases_clean["Order Date"] = pd.to_datetime(df_purchases_clean["Order Date"])

# Convert columns to "int8"
df_purchases_clean["Quantity"] = df_purchases_clean["Quantity"].astype("int8")

df_purchases_clean["ASIN/ISBN (Product Code)"] = df_purchases_clean[
    "ASIN/ISBN (Product Code)"
].astype("category")

### Adding columns


# Extract the year from "Order Date"
df_purchases_clean["Order Date Year"] = df_purchases_clean["Order Date"].dt.year

# Extract the month from "Order Date" and sort it
df_purchases_clean["Order Date Month"] = pd.Categorical(
    df_purchases_clean["Order Date"].dt.strftime("%b"),
    categories=[
        "Jan",
        "Feb",
        "Mar",
        "Apr",
        "May",
        "Jun",
        "Jul",
        "Aug",
        "Sep",
        "Oct",
        "Nov",
        "Dec",
    ],
    ordered=True,
)

# Create "Purchase Total" column
df_purchases_clean["Purchase Total"] = (
    df_purchases_clean["Purchase Price Per Unit"] * df_purchases_clean["Quantity"]
)

# NA values
df_purchases_clean["Quantity"] = df_purchases_clean["Quantity"].fillna(0)

#### Memory usage dataset now


memory = df_purchases_clean.memory_usage(deep=True)  # RAM memory used by each column
total_memory_gb = memory.sum() / (1024**3)  # transform into GB
print(f"Total memory: {total_memory_gb:.2f} GB")

memory_percent = memory / memory.sum() * 100  # RAM memory used by each column in GB
print("\nMemory by column:")
print(memory_percent)

df_purchases_clean["Survey ResponseID"].nunique()

### Create new dataset - page 1

selected_columns = [
    "Survey ResponseID",
    "Order Date",
    "Order Date Year",
    "Order Date Month",
    "Category",
    "Quantity",
    "Purchase Price Per Unit",
    "Purchase Total",
    "ASIN/ISBN (Product Code)",
]
df_purchases_clean = df_purchases_clean[selected_columns]

### Save dataset


df_purchases_clean.to_parquet("data/ready/amazon_purchases.parquet")  # parquet format
