==> data/processing/02_data_processing_survey.ipynb <==
## Dataset: survey

import pandas as pd

### Import dataset

df_survey_clean = pd.read_csv("data/raw/survey.csv")

### Exploratory data analysis

df_survey_clean.info()

df_survey_clean["Survey ResponseID"].nunique()

df_survey_clean["Q-sexual-orientation"].value_counts()

### Data cleaning

# Edit "Q-demos-gender" column
def replace_gender(gender):
    if gender in ["Prefer not to say", "Other"]:
        return "Other or prefer not to say"
    return gender


# Aplicando a função à coluna
df_survey_clean["Q-demos-gender"] = df_survey_clean["Q-demos-gender"].apply(
    replace_gender
)


# Edit "Q-demos-age" column
def clean_age_category(age):
    if "and older" in age:
        return "65+"
    return age.replace(" years", "")


df_survey_clean["Q-demos-age"] = df_survey_clean["Q-demos-age"].apply(
    clean_age_category
)


# Edit "Q-demos-education" column
def remove_parentheses(text):
    return text.split("(")[0].strip()


df_survey_clean["Q-demos-education"] = df_survey_clean["Q-demos-education"].apply(
    remove_parentheses
)


# Edit "Q-demos-income" column
def clean_income(value):
    if "Less than" in value:
        return "Under $25K"
    elif "or more" in value:
        return "Over $150K"
    elif "Prefer not to say" in value:
        return value
    else:
        value = value.replace("$", "").replace(",", "").replace(" - ", "-")
        min_val, max_val = value.split("-")
        min_val = int(min_val) // 1000
        max_val = int(max_val) // 1000
        return f"${min_val} - ${max_val}.9K"


df_survey_clean["Q-demos-income"] = df_survey_clean["Q-demos-income"].apply(
    clean_income
)

# Edit "Q-sexual-orientation" column
df_survey_clean["Q-sexual-orientation"] = df_survey_clean["Q-sexual-orientation"].apply(
    remove_parentheses
)


def capitalized_case(s):
    if pd.isna(s):
        return s
    words = s.split("_")
    return " ".join([words[0].capitalize()] + [word.lower() for word in words[1:]])


df_survey_clean["Q-sexual-orientation"] = df_survey_clean["Q-sexual-orientation"].apply(
    capitalized_case
)

# Edit "Q-amazon-use-howmany" column
df_survey_clean["Q-amazon-use-howmany"] = df_survey_clean["Q-amazon-use-howmany"].apply(
    remove_parentheses
)

# Edit "Q-amazon-use-hh-size" column
df_survey_clean["Q-amazon-use-hh-size"] = df_survey_clean["Q-amazon-use-hh-size"].apply(
    remove_parentheses
)


# Edit "Q-amazon-use-how-oft" column
def clean_frequency(value):
    if "Less than" in value:
        return "< 5 times/month"
    elif "More than" in value:
        return "> 10 times/month"
    else:
        return value.replace("times per month", "times/month")


df_survey_clean["Q-amazon-use-how-oft"] = df_survey_clean["Q-amazon-use-how-oft"].apply(
    clean_frequency
)

# Edit "Q-demos-state" column
df_survey_clean["Q-demos-state"] = df_survey_clean["Q-demos-state"].replace(
    "I did not reside in the United States", "No information"
)

df_survey_clean["Q-demos-state"].value_counts().sort_index()

### Merge dataset: Survey + US Region

# Import region state dataset
df_census = pd.read_csv("data/raw/us_census_bureau_regions_divisions.csv")

df_census.info()

### Data cleaning

# Add new row "No information"
df_census.loc[len(df_census)] = ["No information"] * 4

# rename column
df_census = df_census.rename(columns={"State": "Q-demos-state"})

### Merge dataset

df_survey = pd.merge(df_census, df_survey_clean, on="Q-demos-state", how="inner")

### Reduce dataset size

# Convert columns to "category"
df_survey = df_survey.apply(
    lambda col: (
        col.astype("category")
        if col.dtype == "object" and col.name != "Survey ResponseID"
        else col
    )
)

df_survey["Survey ResponseID"].nunique()

### Reorder columns

df_survey.columns.tolist()

new_order = [
    "Survey ResponseID",
    "Q-demos-state",
    "State Code",
    "Region",
    "Division",
    "Q-demos-age",
    "Q-demos-hispanic",
    "Q-demos-race",
    "Q-demos-education",
    "Q-demos-income",
    "Q-demos-gender",
    "Q-sexual-orientation",
    "Q-amazon-use-howmany",
    "Q-amazon-use-hh-size",
    "Q-amazon-use-how-oft",
    "Q-substance-use-cigarettes",
    "Q-substance-use-marijuana",
    "Q-substance-use-alcohol",
    "Q-personal-diabetes",
    "Q-personal-wheelchair",
    "Q-life-changes",
    "Q-sell-YOUR-data",
    "Q-sell-consumer-data",
    "Q-small-biz-use",
    "Q-census-use",
    "Q-research-society",
]

df_survey = df_survey[new_order]

df_survey.head(2)

df_survey["Q-demos-gender"].value_counts()

df_survey.info()

### Save dataset

df_survey.to_parquet("data/ready/amazon_survey.parquet")
